+++ src/components/VectormapPanel.tsx
@@ -1487,7 +1487,7 @@ export const VectormapPanel: React.FC<Props> = ({
         onToggleGroup={handleToggleGroup}
         activeLabelView={labelView}
         onSelectLabelView={handleSelectLabelView}
-        layerOrder={options.layerOrder ?? { groupOrder: [], itemOrder: [] }}
+        layerOrder={options.layerOrder ?? { groupOrder: [], itemOrder: [], collapsedGroups: [] }}
       />
       <BasemapControl
         choices={basemapChoices}